+++ R/class_settings.R
@@ -148,7 +148,10 @@ settings_validate <- function(settings) {
     settings$dimensions
   )
   tar_assert_chr(settings$iteration)
-  tar_assert_in(settings$error, c("stop", "continue", "abridge", "workspace"))
+  tar_assert_in(
+    settings$error,
+    c("stop", "continue", "abridge", "workspace", "null")
+  )
   tar_assert_in(settings$memory, c("persistent", "transient"))
   tar_assert_lgl(settings$garbage_collection)
   tar_assert_scalar(settings$garbage_collection)